Revised error handling
In VectorMaths.jl, define new error types ColError and DataError with
formatted error messages that can be modified, when thrown.
Use throw instead of error with new error types.
Outsource code for function vector_maths to vector_mean_and_sum.jl and
include in VectorMaths.
In function get_headers, omit else statement with error for vectors when
checking output as this is already check before and never reached here.
Fix error, when x and y columns are assigned with vectors of different
length. Use DataError instead of skipping the DataFrame with an error
message. Otherwise renaming the DataFrame headers will not work due to
different length of colheaders and actual columns. Throw errors for
xcols of length 1 if different from ycols, otherwise, xdata is not
duplicated correctly.
Updated packages in Manifest.toml.